+++ src/OWML.Common/Interfaces/IModVersionChecker.cs
@@ -1,0 +1,7 @@
+﻿namespace OWML.Common
+{
+	public interface IModVersionChecker 
+	{
+		bool CheckModVersion(IModData data);
+	}
+}
