+++ src/OWML.ModLoader/Owo.cs
@@ -28,7 +28,8 @@ namespace OWML.ModLoader
 		private readonly IApplicationHelper _appHelper;
 		private readonly IProcessHelper _processHelper;
 		private readonly IModUnityEvents _unityEvents;
-		private IModManifest _owmlManifest;
+		private readonly IModManifest _owmlManifest;
+		private readonly IModVersionChecker _modVersionChecker;
 		private readonly IList<IModBehaviour> _modList = new List<IModBehaviour>();
 
 		public Owo(
@@ -43,7 +44,8 @@ namespace OWML.ModLoader
 			IGameObjectHelper goHelper,
 			IApplicationHelper appHelper,
 			IProcessHelper processHelper,
-			IModUnityEvents unityEvents)
+			IModUnityEvents unityEvents,
+			IModVersionChecker modVersionChecker)
 		{
 			_modFinder = modFinder;
 			_console = console;
@@ -57,6 +59,7 @@ namespace OWML.ModLoader
 			_appHelper = appHelper;
 			_processHelper = processHelper;
 			_unityEvents = unityEvents;
+			_modVersionChecker = modVersionChecker;
 			_owmlManifest = JsonHelper.LoadJsonObject<ModManifest>($"{_owmlConfig.ManagedPath}/{Constants.OwmlManifestFileName}");
 		}
 
@@ -88,7 +91,7 @@ namespace OWML.ModLoader
 
 				missingDependencies.ForEach(dependency => _console.WriteLine($"Error! {modData.Manifest.UniqueName} needs {dependency}, but it's disabled/missing!", MessageType.Error));
 
-				var shouldLoad = CheckModVersion(modData);
+				var shouldLoad = _modVersionChecker.CheckModVersion(modData);
 				if (!shouldLoad)
 				{
 					continue;
@@ -106,80 +109,6 @@ namespace OWML.ModLoader
 				_menus.ModsMenu?.AddMod(modData, initMod);
 				_modList.Add(initMod);
 			}
-		}
-
-		private bool CheckModVersion(IModData data)
-		{
-			if (_owmlManifest == null)
-			{
-				_console.WriteLine("OWML Manifest is null.", MessageType.Error);
-				return true;
-			}
-
-			var owmlVersion = _owmlManifest.Version;
-
-			(int, int, int) SplitIntoInts(string version, string modname)
-			{
-				if (string.IsNullOrEmpty(version))
-				{
-					_console.WriteLine($"Could not read OWML version of \"{modname}\" - No version found. This mod will not be loaded.", MessageType.Error);
-					return (-1, -1, -1);
-				}
-
-				var split = version.Split('.');
-				if (split.Length < 3)
-				{
-					_console.WriteLine($"Could not read OWML version of \"{modname}\" - Less than 3 digits. This mod will not be loaded.", MessageType.Error);
-					return (-1, -1, -1);
-				}
-
-				var success = true;
-				success &= int.TryParse(split[0], out int major);
-				success &= int.TryParse(split[1], out int minor);
-				success &= int.TryParse(split[2], out int patch);
-				if (!success)
-				{
-					_console.WriteLine($"Could not read OWML version of \"{modname}\" - Could not parse as digits. This mod will not be loaded.", MessageType.Error);
-					return (-1, -1, -1);
-				}
-
-				return (major, minor, patch);
-			}
-
-			var splitOwmlVersion = SplitIntoInts(owmlVersion, "OWML");
-			var splitModVersion = SplitIntoInts(data.Manifest.OWMLVersion, data.Manifest.UniqueName);
-
-			var (owmlMajor, owmlMinor) = (splitOwmlVersion.Item1, splitOwmlVersion.Item2);
-			var (modMajor, modMinor) = (splitModVersion.Item1, splitModVersion.Item2);
-
-			if (splitOwmlVersion == (-1, -1, -1)|| splitModVersion == (-1, -1, -1))
-			{
-				return false;
-			}
-
-			var mismatchText = $"Mismatch between OWML version expected by {data.Manifest.UniqueName} and installed OWML version." +
-					$"\r\nOWML version expected by {data.Manifest.UniqueName} : {data.Manifest.OWMLVersion}" +
-					$"\r\nOWML version installed : {owmlVersion}\r\n";
-
-			if (owmlMajor != modMajor)
-			{
-				_console.WriteLine(mismatchText + $"As the mismatch affects X.~.~, this mod will not be loaded.", MessageType.Error);
-				return false;
-			}
-
-			if (owmlMinor < modMinor)
-			{
-				_console.WriteLine(mismatchText + $"As the the mismatch affects ~.X.~, and the OWML version is lower, the mod will not be loaded.", MessageType.Error);
-				return false;
-			}
-
-			if (owmlMinor > modMinor)
-			{
-				_console.WriteLine(mismatchText + $"As the the mismatch affects ~.X.~, and the OWML version is higher, the mod will still load.", MessageType.Warning);
-				return true;
-			}
-
-			return true;
 		}
 
 		private IEnumerable<IModData> SortMods(IList<IModData> mods)
